PC-78 Create Header.jsx component and style it
