+++ app/assets/stylesheets/application.bootstrap.scss
@@ -21,3 +21,4 @@ $utilities: map-merge(
 @import 'icons';
 @import 'buttons';
 @import 'customer_info';
+@import 'application/pc-header';
