+++ app/assets/stylesheets/application/pc-header.scss
@@ -1,0 +1,18 @@
+.pc-header {
+  height: 60px;
+
+  &-logo {
+    width: 100px;
+    height: 25px;
+  }
+
+  //&-padding {
+  //  &-left {
+  //    padding-left: 160px;
+  //  }
+  //
+  //  &-right {
+  //    padding-right: 144px;
+  //  }
+  //}
+}
