+++ app/javascript/components/layout/Layout.jsx
@@ -1,13 +1,11 @@
 import React from 'react'
-import { NavbarTop } from './NavbarTop'
+import { Header } from './Header'
 import { Outlet } from 'react-router-dom'
 
 export const Layout = () => {
   return (
     <div className={'d-flex flex-column min-vh-100'}>
-      <header>
-        <NavbarTop />
-      </header>
+      <Header />
       <main>
         <Outlet />
       </main>
