+++ app/javascript/components/layout/index.js
@@ -1,1 +1,2 @@
 export * from "./Layout"
+export * from "./Header"
